Organized App.js there are only 3 components such as Navbar, MainMenu and Footer All of the main contents are now in MainMenu component
